LayoutTests/platform/wincairo/accessibility/aria-selected-menu-items-expected.txt

This tests that focused/active menu items are reported as selected children of the parent container.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


First menu has 0 selected child(ren)
Second menu has 0 selected child(ren)
First menubar has 0 selected child(ren)
Second menubar has 0 selected child(ren)
PASS successfullyParsed is true

TEST COMPLETE
